ref: extract api logic to controller layer
